fix #712 - Saving Project does not Convert *.resources to *.resx for use in Visual Studio 2015 on C# & VB Projects.
